+++ Api/AdvertisementApi/AdIntegration.Repository/Repositories/Entity Repositories/CommentRepository.cs
@@ -1,6 +1,7 @@
 ﻿using AdIntegration.Data.DatabaseContext;
 using AdIntegration.Data.Entities;
 using AdIntegration.Repository.Interfaces;
+using Microsoft.EntityFrameworkCore;
 
 namespace AdIntegration.Repository.Repositories;
 
@@ -12,48 +13,66 @@ public class CommentRepository : ICommentRepository
         _context = context;
     }
 
-    public Comment CreateComment(Comment comment)
+    public async Task<Comment> CreateComment(Comment comment)
     {
         _context.Comments.Add(comment);
-        _context.SaveChanges();
+        await _context.SaveChangesAsync();
         return comment;
     }
 
-    public Comment DeleteCommentById(int id)
+    public async Task<Comment> DeleteCommentById(int id)
     {
-        var deleteComment = GetCommentById(id);
-        _context.Comments.Remove(deleteComment);
-        _context.SaveChanges();
+        var deleteComment = await GetCommentById(id);
+
+        if (deleteComment != null)
+        {
+            _context.Comments.Remove(deleteComment);
+            await _context.SaveChangesAsync();
+        }
 
         return deleteComment;
     }
 
-    public IEnumerable<Comment> GetAllComments()
+    public async Task<IEnumerable<Comment>> GetAllComments()
     {
-        var comments = _context.Comments.ToList();
+        var comments = await _context.Comments.ToListAsync();
         return comments;
     }
 
-    public Comment UpdateComment(int id, Comment comment)
+    public async Task<Comment> UpdateComment(int id, Comment comment)
     {
-        var foundComment = _context.Comments.Find(id);
+        var foundComment = await _context.Comments.FindAsync(id);
+
         if (foundComment != null)
         {
             _context.Comments.Update(foundComment);
-            _context.SaveChanges();
-            return foundComment;
+            await _context.SaveChangesAsync();
         }
-        return null;
+
+        return foundComment;
     }
 
-    public Comment GetCommentById(int id)
+    public async Task<Comment> GetCommentById(int id)
     {
-        var comment = _context.Comments.Find(id);
+        var comment = await _context.Comments.FindAsync(id);
+        return comment;
+    }
 
-        if (comment != null)
-        {
-            return comment;
-        }
-        return null;
+    public async Task<IEnumerable<Comment>> GetAllCommentsFromUser(SystemUser user)
+    {
+        var comments = await _context.Comments
+            .Where(comment => comment.UserId == user.UserId)
+            .ToListAsync();
+
+        return comments;
+    }
+
+    public async Task<IEnumerable<Comment>> GetAllCommentsFromPost(Post post)
+    {
+        var comments = await _context.Comments
+            .Where(comment => comment.Id == post.Id)
+            .ToListAsync();
+
+        return comments;
     }
 }
